got app build and containerized.
Tested and it works :)

Next step is to connect "Web -> App" and make it easier for the consumer
to hit the web endpoint.

It would also be nice if the developer could add breakpoints, and it
would hit on the container. I know what Visual Studio 2017 is able to do
this - not sure how, but it does seem possible.